+++ client/scripts/components/doWorkout/workoutTypes/doCustom.js
@@ -1,29 +1,29 @@
 //Not using since 9/4/15
 
-'use strict';
+// 'use strict';
 
-var React = require('react-native');
-var doWorkoutActions = require('../../../actions/doWorkoutActions');
+// var React = require('react-native');
+// var doWorkoutActions = require('../../../actions/doWorkoutActions');
 
-var {
-  StyleSheet,
-  Text,
-  View,
-} = React;
+// var {
+//   StyleSheet,
+//   Text,
+//   View,
+// } = React;
 
-var CustomWorkout = React.createClass({
+// var CustomWorkout = React.createClass({
 
-  render: function(){
-    var instructions = this.props.workout.instructions;
+//   render: function(){
+//     var instructions = this.props.workout.instructions;
 
-    return (
-      /* jshint ignore:start */
-      <View>
-        <Text>{instructions}</Text>
-      </View>
-      /* jshint ignore:end */
-    );
-  }
-});
+//     return (
+//       /* jshint ignore:start */
+//       <View>
+//         <Text>{instructions}</Text>
+//       </View>
+//       /* jshint ignore:end */
+//     );
+//   }
+// });
 
-module.exports = CustomWorkout;
+// module.exports = CustomWorkout;
